Completed the hacs_uart module.
The hacs_uart module can continusouly receive data from USART RX,
and notifies the user of this module upon half-transfer and
transfer-complete. The user should provide two callbacks corresponding
to the HT and TC as mentioned above.

To parse data, it is recommended that user create a state machine which
takes the byte stream from the hacs_uart module as input.