Added BMP085 driver and other changes
Added the BMP085 driver. User of this driver can request a singleshot
sample and gets notified via a callback when the conversion is complete.
This is the first driver that follows a rigorous error-checking style.

Introduced the HACS_REQUIRES macro to reduce the pain of error handling.

Moved driver tasks creation to each driver's init function, instead of
in the app init routine.