Reserved first two sectors of flash for persistent storage
Shifted code placement to start from the third flash sector. Modified
the .icf file and set VTOR on boot to relocate vector table.

Flash sector 0 and 1 are to be used by the future persistent storage
implementation. The impl will use the double-banking method. The storage
will mainly be used to store calibration information