Updated MAVLink to include mag cal result
When a MagCalResult message is received, the mag calibration information
is saved to RAM, and applied to every magnetometer reading.

Fixed a bug in the ground station radio micro that causes ACK payload to
always have length +1.

TODO: implement  persistent storage. Save cal info to flash and read it
during boot.