Added HACS bootloader
The bootloader is located at the first flash sector. Right now all it
does is jumping to the HACS app. In this way, I can reserve flash sector
1 and 2 for env storage, as they won't be overwritten when a new app
image is loaded.

Note that the app address is hardcoded in bootloader_main.c. It is
assumed that the .icf file used by the app also follow the same address.