Implemented the watchdog microcontroller firmware
The main function of the watchdog micro is to control whether the servo
outputs should be driven by the ST or by the RC receiver directly.

The watchdog micro constantly monitors the "checkin" signal from the ST.
The signal should be always changing to indicate that the ST is alive.
If the signal stop changing, the watchdog micro will force the RC mux
into bypass mode, routing all the receiver channels to the servo outputs
directly.

The watchdog micro also monitors receiver_chan_6, so that the pilot can
control the RC mux. Note that in the case that the ST dies, pilot
control of the mux will be overriden, and only bypass mode can be
active.