Fixed a RC receiver bug
It turns out that my assumption that "there is a FIRST channel" is
wrong. On some receivers, the usually 'first' channel is not always
'first'. In my case, when the aileron channel has maximum width, its
rising edge actually precedes that of the throttle channel (But in most
of the cases, the rising edge of the aileron channel is after that of
the throttle channel).

Another implicit assumption was broken: It's not always the case that
the pulse of each channel follow one another without overlapping. On the
receiver that I use, these pulses actually overlaps in time!!