Initial version of the System Identification mode
Support 4 System ID modes

1. Free running. AILE, ELEV and RUDD are all controlled by the pilot.
The system will record these three channels and the resulting states of
the aircraft

2. AILE/ELEV/RUDD mode. Only one control surface is active. HACS will
generate a frequency-sweep signal for 10 seconds. There are two
frequency ranges: 0.1~4 Hz and 3~10 Hz.

I also made a lot of other changes in order for the system id to
function correctly. The actuator API is again changed. The telemetry
module is also changed to support systemId-related messages.