Bug fixes (race condition)
1. Added critical sections to multiple places where potential race
condition may be a problem. Hope this will fix the random RC death
problem (When a RC channel is dead, that GPIO's corresponding bit in the
EXTI->RTSR/FTSR is not set, which makes no sense!!)

2. Changed the way how the RC receiver module determines a sampling
cycle has completed. No longer rely on the notion of a "last channel". A
sampling cycle is considered completed when all the channels have been
sampled.

3. When the system identification experiment starts, I need to remember
the trim values. However, do NOT write to PSTORE at that point, because
the FLASH programming stuff will mess up with the RC readings...

4. Let the sysid module report trim values prior to sending experiment
data